Add files via upload
I decides to add a new game mode called "bullseye". In this mode the program generates a random number, and the user uses the slider to get it as close to the number as they can. After every try they have the option to start over or to do a new round. I was having problems with the start over button because not all the values were returning to zero. I used youtube videos to help me with the problem.